requirements.txt
patchelf==0.12
numba==0.57.0
class-resolver==0.3.10
numpy==1.23.0
imageio==2.29.0
matplotlib==3.7.1
hydra-core==1.3.2
open3d==0.13.0
pandas==2.0.1
plotly==5.14.1
pytorch3d==0.3.0
scipy==1.10.1
tensorboard==2.13.0
tqdm==4.65.0
transforms3d==0.4.1
wandb==0.15.3
cython==0.29.26
h5py==3.7.0
lxml==4.9.1
opencv-python==4.6.0.66
gym==0.21.0
mani-skill2==0.4.2
mujoco-py
pydrive
numpy-quaternion
-e third_party/robomimic-0.2.0 
-e third_party/rrl-dependencies
-e third_party/rrl-dependencies/mj_envs
-e third_party/rrl-dependencies/mjrl
-e third_party/robosuite
-e third_party/stable-baselines3